feat(document): Remove legacy document handling logic
Deleted obsolete files related to an old implementation of document handling in the LangAssembler project. This includes classes and interfaces related to documents, lines, and enumerations. The new implementation does not require these files, which makes the project cleaner and easier to maintain.

+++ src/LangAssembler.Lexer/Extensions/LexerExtensions.cs
@@ -10,7 +10,7 @@ public static class LexerExtensions
     /// </summary>
     /// <param name="lexer">The instance of ILexer to retrieve the farthest index from.</param>
     /// <returns>Returns the farthest index a previous match reached, or the current lexer position.</returns>
-    public static int GetFarthestIndexed(this ILexer lexer) => 
+    public static long GetFarthestIndexed(this ILexer lexer) => 
         lexer.PreviousMatches.LastOrDefault() is {} last 
             ? last.TokenStart + last.TokenLength() 
             : lexer.Position;
